refactor: 메서드 시그니처 변경
