+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -2,10 +2,9 @@ package org.apache.coyote.http11;
 
 import com.techcourse.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.dto.HttpRequest;
-import org.apache.coyote.http11.handler.Handler;
-import org.apache.coyote.http11.util.HttpRequestParser;
-import org.apache.coyote.http11.helper.Responses;
+import org.apache.coyote.http11.request.dto.HttpRequest;
+import org.apache.coyote.http11.handler.DispatcherHandler;
+import org.apache.coyote.http11.request.util.HttpRequestParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,9 +16,9 @@ public class Http11Processor implements Runnable, Processor {
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
 
     private final Socket connection;
-    private final Handler dispatcher;
+    private final DispatcherHandler dispatcher;
 
-    public Http11Processor(Socket connection, Handler dispatcher) {
+    public Http11Processor(Socket connection, DispatcherHandler dispatcher) {
         this.connection = connection;
         this.dispatcher = dispatcher;
     }
@@ -35,14 +34,9 @@ public class Http11Processor implements Runnable, Processor {
         try (final var inputStream = connection.getInputStream();
              final var outputStream = connection.getOutputStream()) {
             HttpRequest httpRequest = HttpRequestParser.parse(inputStream);
-            dispatcher.handle(httpRequest, outputStream);
+            dispatcher.dispatch(httpRequest, outputStream);
         } catch (IOException | UncheckedServletException e) {
             log.error(e.getMessage(), e);
-            try {
-                Responses.serverError(connection.getOutputStream(), "HTTP/1.1");
-            } catch (IOException ioException) {
-                log.error("서버 에러 응답 전송 실패: {}", ioException.getMessage(), ioException);
-            }
         }
     }
 }
