+++ tomcat/src/main/java/org/apache/coyote/http11/handler/Handler.java
@@ -1,12 +1,12 @@
 package org.apache.coyote.http11.handler;
 
 import java.io.IOException;
-import java.io.OutputStream;
-import org.apache.coyote.http11.dto.HttpRequest;
+import org.apache.coyote.http11.request.dto.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public interface Handler {
 
     boolean canHandle(HttpRequest request);
 
-    void handle(HttpRequest request, OutputStream outputStream) throws IOException;
+    void handle(HttpRequest request, HttpResponse response) throws IOException;
 }
